Log enhancement and bug fix
1. Replace the poor log mechanism with python logging module. Now all
important state are logged and page source saved when necessary. This
helps a lot when debugging issues.

2. Fixed bug on parsing staff list and investor list. Now there will be
less exception during extract_info. If the webpage does not has the
desire element, it will return null.

3. Tackled a very strange issue with Selenium get(url) method. It seems
Selenium will stuck when the requested url has robot detection on. I've
searched almost all the possible solution that I can think of, still
cannot figure out the root cause for this, even if I configured the
timeout of webpage loading. So I have to isolate this atom action into a
child thread of scraper, and join with in using a timeout. So that when
Selenium is dead again, I can terminate scraper thread and monitor
thread.